feat: Add wgpu and draw color in window
